backend/requirements.txt
Django==4.2.20
djangorestframework==3.15.2
opencv_python==4.7.0.68
